server/requirements.txt
blinker==1.9.0
click==8.1.8
Deprecated==1.2.18
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Limiter==3.10.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Principal==0.4.0
Flask-Security==5.6.0
Flask-WTF==1.2.2
idna==3.10
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.5
limits==4.0.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mongoengine==0.29.1
ordered-set==4.1.0
packaging==24.2
passlib==1.7.4
Pygments==2.19.1
PyJWT==2.10.1
pymongo==4.11.1
python-dotenv==1.0.1
rich==13.9.4
setuptools==75.8.2
typing_extensions==4.12.2
Werkzeug==3.1.3
wrapt==1.17.2
WTForms==3.2.1
